Moved contentType method into AbstractHTTPClient Class
